Collected small changes and fixes

+++ src/my_page.h
@@ -29,8 +29,8 @@ struct HyperOneCoeff {
 
 template <class T> class MyPage {
  public:
-  int ndatum;    // total # of stored datums
-  int nchunk;    // total # of stored chunks
+  bigint ndatum;    // total # of stored datums
+  int nchunk;       // total # of stored chunks
   MyPage();
   virtual ~MyPage();
 
@@ -105,7 +105,7 @@ template <class T> class MyPage {
                     // 1 = chunk size exceeded maxchunk
                     // 2 = memory allocation error
 #if defined(_OPENMP)
-  char pad[64];     // to avoid false sharing with multi-threading
+  char pad[64];    // to avoid false sharing with multi-threading
 #endif
   void allocate();
   void deallocate();
